test:Am-338 load/delete function test data serenity
